chore: co-locate api fetcher and its types
